A couple of minor issues with KafkaStream
- The handler could never be nulled out. This meant that when a pump
stopped, the handler didn't get nulled like it should. This would lead
to processing messages on the old handler.
- We need to pause the stream before committing in the case of an error.
After this "error condition" commit, the stream will process the next
message. Pause ensures that we don't process the message. As long
as the commit is successful, a new Channel gets created which will
then resume the streawm.