Change default behavior of the auto.offset.reset setting for consumers
Usher had changed this to "smallest" the actual kafka default is "largest".
Originally we set to "smallest" to avoid missing messages after a restart.
However, due to the fact that after the offsets.topic.retention.minutes setting
will completely expire offset messages (by default every day) this causes
a topic without frequent activity to restart at the beginning in the event
of an offset reset. This implicit behavior isn't desired. So we are reverting
to "largest" with the idea that the responsibility is now on the consumer
to "rewind" if need be.